Add 'What is Prebid.js?' and 'How it Works' pages
These pages are designed to give the reader an on-ramp into
understanding Prebid.js before they try to start using it.

They're set up to appear in the site's header nav in the order in which
they should be read.

'How it Works' also includes a diagram showing how Prebid.js fits into
the page, calls external bidders, etc. to supplement the description in
the text.