Code Quality and Style Standards
Define jscs and jshint settings to report on code quality and style standards. Updates code files to standards.
In some cases adapters were also updated, while some jshint options are set to be ignored.
This is not set to enforce quality at commit or build time ... yet. We will want to set that in place soon.
Also, add support for babel and es6 syntax. See the file `test/helpers/pbjs-test-only.js` for an example of
an ES6 module. This file is imported using ES6 syntax in the `test/spec/aliasBidder_spec.js` file and required
the ES5 way in `test/spec/adUnits_spec.js`.