AOL adapter rendering 'undefined' below creative
The response from AOL usually contains the creative code, along with
some code for a tracking pixel.
However, after a user has been synced with ADTECHUS via their tracking
pixel, the subsequent bid responses from AOL will no longer contain
the pixel code.
_addBid() stores the bid response and ad code. Whatever is returned from
response.getCreative() is concatenated with whatever is returned from
response.getPixels() and this is stored as the ad code.
After a user is synced and no longer receiving the tracking pixels,
repsonse.getPixels() will return 'undefined', which is concatenated to
the creative code and rendered to the page.
This pull request changes the AOL adapter to only concatenate the pixel
if it is not 'undefined'.