Pass Deal IDs on bids to ad server as targeting data
If a bid response contains a dealId, its ad server targeting will receive a key with the form `hb_deal_<bidderCode>`. Bids with the highest cpm still always win whether they contain a deal or not. If a winning bid does have a dealId, its ad server targeting will also have a key of `hb_deal` with the value of the deal id. Deals are always sent to the server, regardless of whether `enableSendAllBids` is on or off.

* Construct required key/value pairs for bids with deals
* Always attach bids with deals to targeting
* Test addBidResponse places dealIds in adserver targeting