getLoadTimeDistribution-sorting-fix (#510)
Currently, the load time events are passed through to GA as follows:

0-200ms
200-300ms
300-400ms
400-500ms
500-600ms
600-800ms
800-1000ms
1000-1200ms
1200-1500ms
1500-2000ms
2000ms above

Unfortunately, this means they are sorted like this in the GA sheet:

0-200ms
1000-1200ms
1200-1500ms
1500-2000ms
2000ms above
200-300ms
300-400ms
400-500ms
500-600ms
600-800ms
800-1000ms

This makes the Load Time Distribution chart difficult to read. I've fixed this within my own prebid.js file by adding a leading 0 to the 200-1000ms ranges, e.g. '0200-300ms'. It appears as though in the past the 1000ms+ ranges were entered as follows solving for this issue as well:

t2:1000-1200ms

t2:1200-1500ms

t2:1500-2000ms

t2:2000ms above