Log unsupported ad type only for good bids (#547)
This addresses an issue where bids are returned from the server with a `nobid` field. In that case, bids won't have an ad_type field and an error message is logged that states 'undefined ad type not supported', but bids are still regsitered to the bid manager with status `NO_BID`. This change aims to remove confusion by only loging the error message in cases where a good bid response comes back from the server with an ad_type other than `banner`, which is not currently supported by this adapter.