Modify the adapter API to be able to determine when all bid requests have been responded to
Adapters were free to call addBidResponse multiple times if they wanted to
return multiple bids for a single request.  Due to this, the adUnitBidsBack
and allRequestedBidsBack callbacks were fired at the wrong time.

Adapters are now required to call addBidResponse only once per request, but
it is possible to provide an array of bids.  A new data structure is used to
keep track of the number of addBidResponse calls per ad unit.