Sets a filter for the current bid request as an array of ad unit codes. Applies the filter when calling reporting APIs `getBidResponses`, `getAdserverTargeting`. Queues a bid request if a previous request is still running, closing over instance data. Clears previous bids from placements included in current bid request (remove stale bids).
Note that if a placement is included in one request and a second request as well, and the second request is made before the first request has rendered an ad, the bids for the first request will be cleared and the `renderAd` function will throw an error as the bids cannot be found. However, the placement should be immediately filled with the results of the second request. The error is not caught so as to warn that a placement was overwritten, and care should be taken not to request bids for a placement with a request already in flight.

fixes #772
fixes #590