Filter bid arrays using adUnitsFilter function (#750)
* Filter bid arrays using adUnitsFilter function

The requested and received bid arrays must retain bid objects from multiple calls to `pbjs
.requestBids`. However this causes unexpected behavior in the reporting APIs as bids from
previous auctions were returned. This PR addresses this by creating a `pbjs._adUnitCodes` primary
 data structure to be used as a filter when Prebid needs only the current "auction". While not
 ideal this allows us to retain bid data for in-flight calls to an ad server while also allowing
 additional bid requests to be made without the introduction of Prebid Auction instances.

 Closes #590

* reset externalCallbacks.all.called
* use adUnitsFilter to filter bid arrays to current auction

* targeting function use adUnitsFilter, rollback style fixes

* fix clearPlacements bug

* filter getBidLandscapeTargeting for adUnitCodes

* updates to address review notes

* rebase onto master

* rebase cleanup

* debug shuffle bidders unit test

* restore shuffle bidders test and resetAuction

* use adUnitCodes filter in getWinningBids

* add tests

* revert style changes

* restore webpack settings

* Sets a filter for the current bid request as an array of ad unit codes.
Applies the filter when calling reporting APIs `getBidResponses`, `getAdserverTargeting`.
Queues a bid request if a previous request is still running, closing over instance data.
Clears previous bids from placements included in current bid request (remove stale bids).

Note that if a placement is included in one request and a second request as well, and the second request is made before the first request has rendered an ad, the bids for the first request will be cleared and the `renderAd` function will throw an error as the bids cannot be found. However, the placement should be immediately filled with the results of the second request. The error is not caught so as to warn that a placement was overwritten, and care should be taken not to request bids for a placement with a request already in flight.

fixes #772
fixes #590

* remove line break