Remove doCallbacksIfNeeded
Removes doCallbacksIfNeeded because separating it from
addBidResponse doesn't make a lot of sense. The function name is not
particularly meaningful (doCallbacksIfTimeout would be slightly better).

It also leads to strange ordering issues. For example on video bids that
don't use the prebid cache, timeouts were previously checked after
the bids were added, but for all other bids, the timeouts were checked
before the bids were added. It seems best to just have a single policy
enforced by doing the check when the bid is added.