WIP: call doCallbacks from addBidToAuction
In order to ensure the calling order is correct, call doCallbacks from within
addBidToAuction.

XXX: This breaks the test that's removed in this patch. The reason it breaks is
that the TIMEOUT event is not emitted if all bids have been returned before the
timeout handler is called. Previously the bid manager was rejecting the bid if
the bid triggered the timeout handler, which doesn't really make much sense.
The test SHOULD work if there are two bidders, as the first response should
trigger the timeout. I had a hard time adding a second bidder to the test code,
so I was hoping to get some help.