Update Yieldlab adapter and add official maintainer (#2231)
* Set time to live to 5 minutes

Yieldlab auctions are valid for 5 minutes and not 10. A delivery of an adtag would lead to a new auction after the 5 minute expiration time.

* Use correct identifer for deals

Did stands for demandId which is not a default and not a good identifer for deals. We should use pid (partnershipId) here as a unique identifer for a connection between supply and demand (including deals).

* Set revenue to gross-price

* Add dynamic adsize for vast url

The adsize of a video placement can have a lot of different adsizes and is never "1x1".

* Change maintainer to Yieldlab

Platform Lunar and Yieldlab agree that the Adapter should be offically maintained by Yieldlab.

* Use yieldlab test campaigns

Campaigns set up in Yieldlab's test network to deliver all the time.

* Use yieldlab naming for params

To make it less confusing for customers, we should use the actual naming used in our systems for the params used.

* Update test spec for yieldlab adapter

* Add test for pid to dealId mapping