Bug fix iOS
Local Storage (localStorage or sessionStorage) will be broken in many cases. When it is disabled due to some privacy settings, any attempts to access window.localStorage or setItem() getItem() methods will throw exceptions.

Here are some notable examples where it breaks:
- Mobile Chrome as a webview in many apps (i.e. official Twitter App on Android)
- Chrome and Chrome Mobile with Local storage/cookies disabled (privacy settings).
- Firefox with localStorage disabled.

It is important that using the local storage is made in a try/catch block or Prebid won't work...

It is directly impacting Prebid client monetisation!