s2sTesting: random number moved to global (#3851)
* s2sTesting: random number moved to global

We got a request that if if 10% of bidderA and 10% of bidderB should be server-side, they should be the same 10% of requests.

This is a simple approach to the change that implies that the same page experience only throws one random number for the purposes of s2sTesting. i.e. if there are dynamic ads on the page, all newly created ads on the page will continue to use the same number. This is ok from the purposes of the s2sTesting module.

* Added get for randNum + fix lint & tests

* update to remove getter

* remove log

* update test