Hummingbird analytics adapter prototype
This changeset provides an MVP of a Maven analytics adapter, which
gathers bid/no bid information for each auction on a page and sends it
to an endpoint defined in a config object. Info gathered includes a few
details on the environment (country, pod, screen size, etc.) and, for
submitted bids, CPM and time-to-respond data enabling us to do some
counterfactual exploration of our bidder setup.