PBID-39: Pass a true CCPA optout status to adapters when the EID has been suppressed
The userId/eids module will not consider our ID system for inclusion in the EIDs object if our ID value is not a string.
Unfortunately when we write our cookie without an EID (in the case where CCPA optout is true) then the deserialized EID value is undefined, to save space in the cookie.

So this is a hack that will return an empty string when Prebid is building the EIDs object so that we can still pass our optout status to those that require it to understand why our ID may be missing.