Prebid core & PBS Adapter: debugging tools to intercept bid requests and mock their response
This adds the ability to 'intercept' bid requests with a mock response. An intercepted bid does not go through the normal `buildRequests` -> network -> `interpretResponse` flow; a dummy response is instead direclty returned to the auction as if it originated from `interpretResponse`.

This is mostly transparent to other parts of the stack (validation, events, etc), with the exception of user syncs - their interface requires the bidder's HTTP response, so they won't run if all bids in the request are intercepted (and thus no network activity happens).