Prebid core & currency module: fix race condition on slow fetch of currency conversion file
The currency module hooks into the auction's `addBidResponse` flow to modify bid responses with currency info. To do this it needs to fetch conversion rates from the network, but the auction expects `addBidResponse` to complete synchronously, and things break down if all bidder network calls complete before the currency file fetch.

This fix allows the `addBidResponse` hook to return a promise. If it does, the auction will wait for it before ending.

Addresses https://github.com/prebid/Prebid.js/issues/7765