Mapping table initialization
A constant initialises a mapping table where each possible key used by IAS is mapped to itself.
The dataProviders config for the IAS RTD module accepts an optional new property: keyMappings. The init function reads this property and uses it to overwrite any key that is already in the pre-initialised mapping table with the new value provided.
In a future addition, the mapping table will be used to populate the key-values actually sent to the adserver.