Support IAD SDA test in Permutive module
Updates the Permutive RTD module to support a test for segmentation with
the new IAB standard data taxonomy.

Initially we are mapping a small selection of Permutive to IAB cohort
IDs and including these in the ORTB2 object. We also set the `segtax`
value appropriately for the Index exchange, which we're using for this
test.