getAdUnitPath extracted to manage the adUnitPath option
getAdUnitPath obtains the p parameter using the exact same technique used before. After obtaining it, it checks if there is an adUnitPath configuration parameter, and if it has a non-empty string mapped to a property named like the ad slot id. If and only if this is the case, it overwrites the p parameter with that string.

Since different adservers may have different conventions, no check is run on the formatting of the ad unit path proper.