Provide backwards compatibility for transparency object
When the property `transparency.params` gets passed in the `ortb2` object, the Rubicon adapter throws an error. 

This change will provide backwards compatibility for the `params` property by adding it to the `dsaparams` property.

If `params` or `dsaparams` does not exist, the value will be turned into an empty string. The same is done for the `domain` property. This ensures that the Bid Adapter will not error if these properties do not exist.